0x11: Task3 C source code and some modifications
Added Hexadecimal Conversion Specifiers as a
function including small and capital alphabets.

Added Octal Conversion Specifiers as a function.

Modified the _str_func.c with the lost chunk of
code in the previous commit.

Updated the main header file with the required
new prototypes for Hexadecimal and Octal
Conversion Specifiers.

Created a new subdirectory to include our test
files and ALX test file.